requirements.txt
attrs==21.4.0
autosemver==0.5.5
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
dulwich==0.19.16
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
PyPDF2==1.26.0
pytest==7.1.1
pytest-mock==3.7.0
python-magic==0.4.25
refextract==1.1.3
requests==2.27.1
six==1.16.0
tomli==2.0.1
Unidecode==1.3.4
urllib3==1.26.9
validators==0.18.2
